% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_label.R
\name{has_label}
\alias{has_label}
\title{A selector function to use in \pkg{tidyverse} functions}
\usage{
has_label(labels)
}
\arguments{
\item{labels}{A character vector of labels you want to operate on}
}
\value{
A numeric vector containing the position of the columns with the
requested labels
}
\description{
A selector function to use in \pkg{tidyverse} functions
}
\note{
Using this in a pipeline results in a 'safeframe' object, but does not
maintain the variable labels at this time. It is primarily useful to make
your pipelines human readable.
}
\examples{
## create safeframe
x <- make_safeframe(cars,
  speed = "Miles per hour",
  dist = "Distance in miles"
)
head(x)

if (require(dplyr) && require(magrittr)) {
  x \%>\%
    select(has_label(c("Miles per hour", "Distance in miles"))) \%>\%
    head()
}
}
