% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_variables.R
\name{label_variables}
\alias{label_variables}
\title{Add labels to variables}
\usage{
label_variables(x, labels)
}
\arguments{
\item{x}{a \code{data.frame} or a \code{tibble}, with at least one column}

\item{labels}{A named list with variable names in \code{x} as list names and the
labels as list values. Values set to \code{NULL} remove the label.}
}
\value{
The function returns the original object with an additional \code{"label"}
attribute on each provided variable.
}
\description{
Internal. This function will label pre-defined variables in a
\code{data.frame} by adding a label attribute to the column. This can be used for
one or multiple variables at the same time.
}
\details{
If used several times, the previous label is removed silently.
Only accepts known variables from the provided \code{data.frame}.
}
