% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lost_labels.R
\name{lost_labels}
\alias{lost_labels}
\title{Check for lost labels and throw relevant warning or error}
\usage{
lost_labels(old, new, lost_action)
}
\arguments{
\item{old}{A named list of old labels.}

\item{new}{A named list of new labels.}

\item{lost_action}{A character string specifying the action to take when
labels are lost. Can be "none", "warning", or "error".}
}
\value{
None. Throws a warning or error if labels are lost.
}
\description{
This internal function checks for labels that are present in the old labels
but not in the new labels. If any labels are lost, it throws a warning or
error based on the specified action.
}
\keyword{internal}
