% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lost_labels_action.R
\name{lost_labels_action}
\alias{lost_labels_action}
\alias{get_lost_labels_action}
\title{Check and set behaviour for lost labels}
\usage{
lost_labels_action(action = c("warning", "error", "none"), quiet = FALSE)

get_lost_labels_action()
}
\arguments{
\item{action}{a \code{character} indicating the behaviour to adopt when labelled
variables have been lost: "error" (default) will issue an error; "warning"
will issue a warning; "none" will do nothing}

\item{quiet}{a \code{logical} indicating if a message should be displayed; only
used outside pipelines}
}
\value{
returns \code{NULL}; the option itself is set in \code{options("datatagr")}
}
\description{
This function determines the behaviour to adopt when labelled variables of a
\code{datatagr} are lost for example through subsetting. This is achieved using
\code{options} defined for the \code{datatagr} package.
}
\details{
The errors or warnings generated by datatagr in case of labelled
variable loss has a custom class of \code{datatagr_error} and \code{datatagr_warning}
respectively.
}
\examples{
# reset default - done automatically at package loading
lost_labels_action()

# check current value
get_lost_labels_action()

# change to issue errors when tags are lost
lost_labels_action("error")
get_lost_labels_action()

# change to ignore when tags are lost
lost_labels_action("none")
get_lost_labels_action()

# reset to default: warning
lost_labels_action()

}
