% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_datatagr.R
\name{make_datatagr}
\alias{make_datatagr}
\title{Create a datatagr from a data.frame}
\usage{
make_datatagr(x, ...)
}
\arguments{
\item{x}{a \code{data.frame} or a \code{tibble}}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> A named list with variable
names in \code{x} as list names and the labels as list values. Values set to
\code{NULL} remove the label. When specifying labels, please also see
\code{default_values}.}
}
\value{
The function returns a \code{datatagr} object.
}
\description{
This function converts a \code{data.frame} or a \code{tibble} into a \code{datatagr} object,
where data are labelled and validated. The output will seem to be the same
\code{data.frame}, but \code{datatagr}-aware packages will then be able to
automatically use labelled fields for further data cleaning and analysis.
}
\examples{

x <- make_datatagr(cars,
  speed = "Miles per hour",
  dist = "Distance in miles"
)

## print result - just first few entries
head(x)

## check tags
labels(x)

## Tags can also be passed as a list with the splice operator (!!!)
my_tags <- list(
  speed = "Miles per hour",
  dist = "Distance in miles"
)
new_x <- make_datatagr(cars, !!!my_tags)

## The output is strictly equivalent to the previous one
identical(x, new_x)

}
\seealso{
\itemize{
\item An overview of the \link{datatagr} package
\item \code{\link[=labels]{labels()}}: for a list of labelled variables in a \code{datatagr}
\item \code{\link[=set_labels]{set_labels()}}: for modifying labels
\item \code{\link[=labels_df]{labels_df()}}: for selecting variables by labels
}
}
