% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_safeframe.R
\name{make_safeframe}
\alias{make_safeframe}
\title{Create a safeframe from a data.frame}
\usage{
make_safeframe(x, ...)
}
\arguments{
\item{x}{a \code{data.frame} or a \code{tibble}}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> A named list with variable
names in \code{x} as list names and the labels as list values. Values set to
\code{NULL} remove the label. When specifying labels, please also see
\code{default_values}.}
}
\value{
The function returns a \code{safeframe} object.
}
\description{
This function converts a \code{data.frame} or a \code{tibble} into a \code{safeframe}
object, where data are labelled and validated. The output will seem to be the
same \code{data.frame}, but \code{safeframe}-aware packages will then be able to
automatically use labelled fields for further data cleaning and analysis.
}
\examples{

x <- make_safeframe(cars,
  speed = "Miles per hour",
  dist = "Distance in miles"
)

## print result - just first few entries
head(x)

## check labels
labels(x)

## Labels can also be passed as a list with the splice operator (!!!)
my_labels <- list(
  speed = "Miles per hour",
  dist = "Distance in miles"
)
new_x <- make_safeframe(cars, !!!my_labels)

## The output is strictly equivalent to the previous one
identical(x, new_x)

}
\seealso{
\itemize{
\item An overview of the \link{safeframe} package
\item \code{\link[=labels]{labels()}}: for a list of labelled variables in a \code{safeframe}
\item \code{\link[=set_labels]{set_labels()}}: for modifying labels
\item \code{\link[=labels_df]{labels_df()}}: for selecting variables by labels
}
}
