% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{names<-.safeframe}
\alias{names<-.safeframe}
\title{Rename columns of a safeframe}
\usage{
\method{names}{safeframe}(x) <- value
}
\arguments{
\item{x}{a \code{safeframe} object}

\item{value}{a \code{character} vector to set the new names of the columns of \code{x}}
}
\value{
a \code{safeframe} with new column names
}
\description{
This function can be used to rename the columns a \code{safeframe} (that is,
adjust
variable names).
}
\examples{
## create safeframe
x <- make_safeframe(cars,
  speed = "Miles per hour",
  dist = "Distance in miles"
)
head(x)

## change names
names(x)[1] <- "mph"

## see results: columns have been updated
head(x)

# This also works with using `dplyr::rename()` because it uses names<-()
# under the hood
if (require(dplyr) && require(magrittr)) {
  x <- x \%>\%
    rename(speed = "mph")
  head(x)
  labels(x)
}
}
