% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.datatagr.R
\name{print.datatagr}
\alias{print.datatagr}
\title{Printing method for datatagr objects}
\usage{
\method{print}{datatagr}(x, ...)
}
\arguments{
\item{x}{a \code{datatagr} object}

\item{...}{further arguments to be passed to 'print'}
}
\value{
Invisibly returns the object.
}
\description{
This function prints datatagr objects.
}
\examples{
if (require(outbreaks)) {
  ## dataset we'll create a datatagr from
  measles_hagelloch_1861

  ## create datatagr
  x <- make_datatagr(measles_hagelloch_1861,
    id = "case_ID",
    date_onset = "date_of_prodrome",
    age = "age",
    gender = "gender"
  )

  ## print object - using only the first few entries
  head(x)

  # version with a tibble
  if (require(tibble) && require(magrittr)) {
    measles_hagelloch_1861 \%>\%
      tibble() \%>\%
      make_datatagr(
        id = "case_ID",
        date_onset = "date_of_prodrome",
        age = "age",
        gender = "gender"
      )
  }
}
}
