% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.datatagr.R
\name{select.datatagr}
\alias{select.datatagr}
\title{Subset columns of a datatagr object}
\usage{
\method{select}{datatagr}(.data, ..., tags)
}
\arguments{
\item{.data}{a \code{datatagr} object}

\item{...}{the variables to select, using \code{dplyr} compatible syntax}

\item{tags}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} It is now recommended to
leverage the \code{\link[=has_tag]{has_tag()}} selection helper rather than this argument.}
}
\value{
The function returns a \code{datatagr} with selected columns.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

This function was deprecated to ensure full compatibility with the default
\code{\link[dplyr:select]{dplyr::select()}} methods. The tag selection feature is now possible via the
\code{\link[=has_tag]{has_tag()}} selection helper.
}
\seealso{
\itemize{
\item \code{\link[=tags_df]{tags_df()}} to return a \code{data.frame} of all tagged variables
}
}
\keyword{deprecated}
