% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_tags.R
\name{set_tags}
\alias{set_tags}
\title{Changes tags of a datatagr object}
\usage{
set_tags(x, ..., tag_defaults = list())
}
\arguments{
\item{x}{a \code{data.frame} or a \code{tibble}}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> A series of tags provided as
\code{tag_name = "column_name"}. When specifying tags, please also see
\code{tag_defaults} to specify default values.}

\item{tag_defaults}{a list of default values for the provided tags. Defaults
to \code{list()}, effectively defaulting to NULL values.}
}
\value{
The function returns a \code{datatagr} object.
}
\description{
This function changes the \code{tags} of a \code{datatagr} object, using the same
syntax as the constructor \code{\link[=make_datatagr]{make_datatagr()}}. If some of the default tags are
missing, they will be added to the final object.
}
\examples{

## create a datatagr
x <- make_datatagr(cars, mph = "speed")
tags(x)

## add new tags and fix an existing one
x <- set_tags(x, distance = "dist")
tags(x)

## remove tags by setting them to NULL
old_tags <- tags(x)
x <- set_tags(x, mph = NULL, distance = NULL)
tags(x)

## setting tags providing a list (used to restore old tags here)
x <- set_tags(x, !!!old_tags)
tags(x)
}
\seealso{
\code{\link[=make_datatagr]{make_datatagr()}} to create a \code{datatagr} object
}
