% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square_bracket.R
\name{[.datatagr}
\alias{[.datatagr}
\alias{sub_datatagr}
\alias{[<-.datatagr}
\alias{[[<-.datatagr}
\alias{$<-.datatagr}
\title{Subsetting of datatagr objects}
\usage{
\method{[}{datatagr}(x, i, j, drop = FALSE)

\method{[}{datatagr}(x, i, j) <- value

\method{[[}{datatagr}(x, i, j) <- value

\method{$}{datatagr}(x, name) <- value
}
\arguments{
\item{x}{a \code{datatagr} object}

\item{i}{a vector of \code{integer} or \code{logical} to subset the rows of the
\code{datatagr}}

\item{j}{a vector of \code{character}, \code{integer}, or \code{logical} to subset the
columns of the \code{datatagr}}

\item{drop}{a \code{logical} indicating if, when a single column is selected, the
\code{data.frame} class should be dropped to return a simple vector, in which
case the \code{datatagr} class is lost as well; defaults to \code{FALSE}}

\item{value}{the replacement to be used for the entries identified in \code{x}}

\item{name}{
    a literal character string or a \link[base]{name} (possibly \link[base]{backtick}
    quoted).  For extraction, this is normally (see under
    \sQuote{Environments}) partially matched to the \code{\link[base]{names}}
    of the object.
  }
}
\value{
If no drop is happening, a \code{datatagr}. Otherwise an atomic vector.
}
\description{
The \verb{[]} and \verb{[[]]} operators for \code{datatagr} objects behaves like for regular
\code{data.frame} or \code{tibble}, but check that labelled variables are not lost, and
takes the appropriate action if this is the case (warning, error, or ignore,
depending on the general option set via \code{\link[=lost_labels_action]{lost_labels_action()}}) .
}
\examples{
if (require(dplyr) && require(magrittr)) {
  ## create a datatagr
  x <- cars \%>\%
    make_datatagr(
      mph = "speed",
      distance = "dist"
    ) \%>\%
    mutate(result = if_else(speed > 50, "fast", "slow")) \%>\%
    set_labels(ticket = "result")
  x

  ## dangerous removal of a labelled column setting it to NULL issues a warning
  x[, 1] <- NULL
  x

  x[[2]] <- NULL
  x

  x$age <- NULL
  x
}
}
\seealso{
\itemize{
\item \code{\link[=lost_labels_action]{lost_labels_action()}} to set the behaviour to adopt when labels are
lost through subsetting; default is to issue a warning
\item \code{\link[=get_lost_labels_action]{get_lost_labels_action()}} to check the current the behaviour
}
}
