% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags_defaults.R
\name{tags_defaults}
\alias{tags_defaults}
\title{Generate default tags for a linelist}
\usage{
tags_defaults()
}
\value{
A named \code{list}.
}
\description{
This function returns a named list providing the default tags for a
\code{linelist} object (all default to NULL).
}
\examples{
tags_defaults()

}
