% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags_df.R
\name{tags_df}
\alias{tags_df}
\title{Extract a data.frame of all tagged variables}
\usage{
tags_df(x)
}
\arguments{
\item{x}{a \code{datatagr} object}
}
\value{
A \code{data.frame} of tagged variables.
}
\description{
This function returns a \code{data.frame} of all the tagged variables stored in a
\code{datatagr}. Note that the output is no longer a \code{datatagr}, but a regular
\code{data.frame}.
}
\examples{

x <- make_datatagr(cars,
  mph = "speed",
  distance = "dist"
)

## get a data.frame of all tagged variables
tags_df(x)
}
