% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags_names.R
\name{tags_names}
\alias{tags_names}
\title{Get the list of tag names used in linelist}
\usage{
tags_names()
}
\value{
The function returns a \code{character} vector.
}
\description{
This function returns the a \code{character} of all tag names used to designate
specific variable types in a \code{linelist}.
}
\examples{
tags_names()

}
\seealso{
\code{\link[=tags_defaults]{tags_defaults()}} for a \code{list} of default values of the tags
}
