% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_labels.R
\name{validate_labels}
\alias{validate_labels}
\title{Checks the labels of a datatagr object}
\usage{
validate_labels(x)
}
\arguments{
\item{x}{a \code{datatagr} object}
}
\value{
If checks pass, a \code{datatagr} object; otherwise issues an error.
}
\description{
This function evaluates the validity of the labels of a \code{datatagr} object by
checking that: i) labels are present ii) labels is a \code{list} of \code{character} or
\code{NULL} values.
}
\examples{
## create a valid datatagr
x <- cars |>
  make_datatagr(
    speed = "Miles per hour",
    dist = "Distance in miles"
  )
x

## the below issues an error as datatagr doesn't know any defaults
## note: tryCatch is only used to avoid a genuine error in the example
tryCatch(validate_datatagr(x), error = paste)

## validation requires you to specify the types directly
validate_datatagr(x,
  speed = c("integer", "numeric"),
  dist = "numeric"
)
}
\seealso{
\code{\link[=validate_types]{validate_types()}} to check if labelled variables have
the right classes
}
