% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_tags.R
\name{validate_tags}
\alias{validate_tags}
\title{Checks the tags of a datatagr object}
\usage{
validate_tags(x)
}
\arguments{
\item{x}{a \code{datatagr} object}
}
\value{
If checks pass, a \code{datatagr} object; otherwise issues an error.
}
\description{
This function evaluates the validity of the tags of a \code{datatagr} object by
checking that: i) tags are present ii) tags is a \code{list} of \code{character} iii)
that all default tags are present iv) tagged variables exist.
}
\examples{
if (require(dplyr) && require(magrittr)) {
  ## create a valid datatagr
  x <- cars \%>\%
    make_datatagr(
      mph = "speed",
      distance = "dist"
    )
  x

  ## the below issues an error as datatagr doesn't know any defaults
  ## note: tryCatch is only used to avoid a genuine error in the example
  tryCatch(validate_datatagr(x), error = paste)

  ## validation requires you to specify the types directly
  validate_datatagr(x, ref_types = tags_types(
    mph = c("integer", "numeric"),
    distance = "numeric"
  ))
}
}
\seealso{
\code{\link[=validate_types]{validate_types()}} to check if tagged variables have
the right classes
}
