% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars_labels.R
\name{vars_labels}
\alias{vars_labels}
\title{Internal printing function for variables and labels}
\usage{
vars_labels(vars, labels)
}
\arguments{
\item{vars}{a \code{character} vector of variable names}

\item{labels}{a \code{character} vector of labels}
}
\description{
Internal printing function for variables and labels
}
