% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{labels}
\alias{labels}
\title{Get the list of labels in a datatagr}
\usage{
labels(x, show_null = FALSE)
}
\arguments{
\item{x}{a \code{datatagr} object}

\item{show_null}{a \code{logical} indicating if the complete list of labels,
including \code{NULL} ones, should be returned; if \code{FALSE}, only labels with a
non-NULL value are returned; defaults to \code{FALSE}}
}
\value{
The function returns a named \code{list} where names indicate which column
they correspond to, and values indicate the relevant labels.
}
\description{
This function returns the list of labels identifying specific variable types
in a \code{datatagr} object.
}
\details{
Labels are stored as the \code{label} attribute of the column variable.
}
\examples{

## make a datatagr
x <- make_datatagr(cars, speed = "Miles per hour")

## check non-null labels
labels(x)

## get a list of all labels, including NULL ones
labels(x, TRUE)
}
