% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lost_tags_action.R
\name{lost_tags_action}
\alias{lost_tags_action}
\alias{get_lost_tags_action}
\title{Check and set behaviour for lost tags}
\usage{
lost_tags_action(action = c("warning", "error", "none"), quiet = FALSE, x)

get_lost_tags_action()
}
\arguments{
\item{action}{a \code{character} indicating the behaviour to adopt when tagged
variables have been lost: "error" (default) will issue an error; "warning"
will issue a warning; "none" will do nothing}

\item{quiet}{a \code{logical} indicating if a message should be displayed; only
used outside pipelines}

\item{x}{deprecated}
}
\value{
returns \code{NULL}; the option itself is set in \code{options("linelist")}
}
\description{
This function determines the behaviour to adopt when tagged variables of a
\code{linelist} are lost e.g. through subsetting. This is achieved using \code{options}
defined for the \code{linelist} package.
}
\details{
The errors or warnings generated by linelist in case of tagged
variable loss has a custom class of \code{linelist_error} and \code{linelist_warning}
respectively.
}
\examples{
# reset default - done automatically at package loading
lost_tags_action()

# check current value
get_lost_tags_action()

# change to issue errors when tags are lost
lost_tags_action("error")
get_lost_tags_action()

# change to ignore when tags are lost
lost_tags_action("none")
get_lost_tags_action()

# reset to default: warning
lost_tags_action()

}
