% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_datatagr.R
\name{make_datatagr}
\alias{make_datatagr}
\title{Create a datatagr from a data.frame}
\usage{
make_datatagr(x, ..., tag_defaults = list(), allow_extra = TRUE)
}
\arguments{
\item{x}{a \code{data.frame} or a \code{tibble}}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> A series of tags provided as
\code{tag_name = "column_name"}, where \code{tag_name} indicates any of the known
variables listed in 'Details' and values indicate their name in \code{x}; see
details for a list of known variable types and their expected content}

\item{tag_defaults}{a list of default values for the provided tags. Defaults
to \code{list()}}

\item{allow_extra}{a \code{logical} indicating if additional data tags not
currently recognized by \code{datatagr} should be allowed; if \code{FALSE}, unknown
tags will trigger an error}
}
\value{
The function returns a \code{datatagr} object.
}
\description{
This function converts a \code{data.frame} or a \code{tibble} into a \code{datatagr} object,
where data are tagged. The output will seem to be the same \code{data.frame}, but
\code{datatagr}-aware packages will then be able to automatically use tagged
fields for further data cleaning and analysis.
}
\details{
Known variable types include:
\itemize{
\item \code{id}: a unique case identifier as \code{numeric} or \code{character}
\item \code{date_onset}: date of symptom onset (see below for date formats)
\item \code{date_reporting}: date of case notification (see below for date formats)
\item \code{date_admission}: date of hospital admission (see below for date formats)
\item \code{date_discharge}: date of hospital discharge (see below for date formats)
\item \code{date_outcome}: date of disease outcome (see below for date formats)
\item \code{date_death}: date of death (see below for date formats)
\item \code{gender}: a \code{factor} or \code{character} indicating the gender of the patient
\item \code{age}: a \code{numeric} indicating the age of the patient, in years
\item \code{location}: a \code{factor} or \code{character} indicating the location of the
patient
\item \code{occupation}: a \code{factor} or \code{character} indicating the professional
activity of the patient
\item \code{hcw}: a \code{logical} indicating if the patient is a health care worker
\item \code{outcome}: a \code{factor} or \code{character} indicating the outcome of the disease
(death or survival)

Dates can be provided in the following formats/types:
\item \code{Date} objects (e.g. using \code{as.Date} on a \code{character} with a correct date
format); this is the recommended format
\item \code{POSIXct/POSIXlt} objects (when a finer scale than days is needed)
\item \code{numeric} values, typically indicating the number of days since the first
case
}
}
\examples{

x <- make_datatagr(cars,
  age = "speed",
  distance = "dist"
)

## print result - just first few entries
head(x)

## check tags
tags(x)

## Tags can also be passed as a list with the splice operator (!!!)
my_tags <- list(
  age = "speed",
  distance = "dist"
)
new_x <- make_datatagr(cars, !!!my_tags)

## The output is strictly equivalent to the previous one
identical(x, new_x)

}
\seealso{
\itemize{
\item An overview of the \link{datatagr} package
\item \code{\link[=tags_types]{tags_types()}}: for the associated accepted types/classes
\item \code{\link[=tags]{tags()}}: for a list of tagged variables in a \code{datatagr}
\item \code{\link[=set_tags]{set_tags()}}: for modifying tags
\item \code{\link[=tags_df]{tags_df()}}: for selecting variables by tags
}
}
