% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{names<-.datatagr}
\alias{names<-.datatagr}
\title{Rename columns of a datatagr}
\usage{
\method{names}{datatagr}(x) <- value
}
\arguments{
\item{x}{a \code{datatagr} object}

\item{value}{a \code{character} vector to set the new names of the columns of \code{x}}
}
\value{
a \code{datatagr} with new column names
}
\description{
This function can be used to rename the columns a \code{datatagr} (that is, adjust
variable names).
}
\examples{
## create datatagr
x <- make_datatagr(cars,
  speed = "Miles per hour",
  dist = "Distance in miles"
)
head(x)

## change names
names(x)[1] <- "mph"

## see results: columns have been updated
head(x)

# This also works with using `dplyr::rename()` because it uses names<-()
# under the hood
if (require(dplyr) && require(magrittr)) {
  x <- x \%>\%
    rename(speed = "mph")
  head(x)
  labels(x)
}
}
