% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.datatagr.R
\name{print.datatagr}
\alias{print.datatagr}
\title{Printing method for datatagr objects}
\usage{
\method{print}{datatagr}(x, ...)
}
\arguments{
\item{x}{a \code{datatagr} object}

\item{...}{further arguments to be passed to 'print'}
}
\value{
Invisibly returns the object.
}
\description{
This function prints datatagr objects.
}
\examples{
## create datatagr
x <- make_datatagr(cars,
  mph = "speed",
  distance = "dist"
)

## print object - using only the first few entries
head(x)

# version with a tibble
if (require(tibble) && require(magrittr)) {
  cars \%>\%
    tibble() \%>\%
    make_datatagr(
      mph = "speed",
      distance = "dist"
    )
}
}
