% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_labels.R
\name{set_labels}
\alias{set_labels}
\title{Change labels of a datatagr object}
\usage{
set_labels(x, ...)
}
\arguments{
\item{x}{a \code{data.frame} or a \code{tibble}}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> A named list with variable names in \code{x} as list names and the labels as list values. Values set to \code{NULL} remove the label. When specifying labels, please also see
\code{default_values}.}
}
\value{
The function returns a \code{datatagr} object.
}
\description{
This function changes the \code{labels} of a \code{datatagr} object, using the same
syntax as the constructor \code{\link[=make_datatagr]{make_datatagr()}}.
}
\examples{

## create a datatagr
x <- make_datatagr(cars, speed = "Miles per hour")
labels(x)

## add new labels and fix an existing one
x <- set_labels(x, dist = "Distance")
labels(x)

## remove labels by setting them to NULL
old_labels <- labels(x)
x <- set_labels(x, speed = NULL, dist = NULL)
labels(x)

## setting labels providing a list (used to restore old labels here)
x <- set_labels(x, !!!old_labels)
labels(x)
}
\seealso{
\code{\link[=make_datatagr]{make_datatagr()}} to create a \code{datatagr} object
}
